+++ src/class03/Code04_BiggerThanRightTwice.java
@@ -19,14 +19,19 @@ public class Code04_BiggerThanRightTwice {
 	}
 
 	public static int merge(int[] arr, int L, int m, int r) {
+		// [L....M]   [M+1....R]
+		
 		int ans = 0;
+		// 目前囊括进来的数，是从[M+1, windowR)
 		int windowR = m + 1;
 		for (int i = L; i <= m; i++) {
-			while (windowR <= r && arr[i] > (arr[windowR] << 1)) {
+			while (windowR <= r && arr[i] > (arr[windowR] * 2)) {
 				windowR++;
 			}
 			ans += windowR - m - 1;
 		}
+		
+		
 		int[] help = new int[r - L + 1];
 		int i = 0;
 		int p1 = L;
